Seccion 38 Manejo de Excepciones Tercera parte
