+++ ManejoDeExcepciones2/src/excepciones/OperacionExcepcion.java
@@ -1,8 +1,1 @@
-package excepciones;
-//check exception requiere un proceso para indicar la excepcion
-public class OperacionExcepcion  extends Exception{
 
-    public OperacionExcepcion(String mensaje){
-        super(mensaje);
-    }
-}
